# frontend/src/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "战鸽数据",
    "description": "Zhange Stats · Steam 游玩统计与圈子成员管理",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/skland/status": {
      "get": {
        "tags": [
          "skland",
          "skland"
        ],
        "summary": "Skland Status",
        "operationId": "skland_status_api_skland_status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "include_roles",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Roles"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SklandStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/skland/logs": {
      "get": {
        "tags": [
          "skland",
          "skland"
        ],
        "summary": "Skland Logs",
        "description": "兼容占位：历史列表已弃用；今日状态见 status（读 *_checkin_logs 缓存）。",
        "operationId": "skland_logs_api_skland_logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CheckinLogOut"
                  },
                  "title": "Response Skland Logs Api Skland Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/skland/bind": {
      "post": {
        "tags": [
          "skland",
          "skland"
        ],
        "summary": "Skland Bind",
        "operationId": "skland_bind_api_skland_bind_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SklandBindRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SklandStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "skland",
          "skland"
        ],
        "summary": "Skland Unbind",
        "operationId": "skland_unbind_api_skland_bind_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SklandStatusOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "skland",
          "skland"
        ],
        "summary": "Skland Update Bind",
        "operationId": "skland_update_bind_api_skland_bind_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SklandBindUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SklandStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/skland/bind/password": {
      "post": {
        "tags": [
          "skland",
          "skland"
        ],
        "summary": "Skland Bind Password",
        "operationId": "skland_bind_password_api_skland_bind_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SklandBindPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SklandStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/skland/bind/sms/send": {
      "post": {
        "tags": [
          "skland",
          "skland"
        ],
        "summary": "Skland Bind Sms Send",
        "operationId": "skland_bind_sms_send_api_skland_bind_sms_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SklandBindSmsSendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SklandBindSmsSendResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/skland/bind/sms": {
      "post": {
        "tags": [
          "skland",
          "skland"
        ],
        "summary": "Skland Bind Sms",
        "operationId": "skland_bind_sms_api_skland_bind_sms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SklandBindSmsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SklandStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/skland/qr/start": {
      "post": {
        "tags": [
          "skland",
          "skland"
        ],
        "summary": "Skland Qr Start",
        "operationId": "skland_qr_start_api_skland_qr_start_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SklandQrStartResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/skland/qr/poll": {
      "get": {
        "tags": [
          "skland",
          "skland"
        ],
        "summary": "Skland Qr Poll",
        "operationId": "skland_qr_poll_api_skland_qr_poll_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scan_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 4,
              "maxLength": 128,
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SklandQrPollResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/skland/checkin": {
      "post": {
        "tags": [
          "skland",
          "skland"
        ],
        "summary": "Skland Checkin Now",
        "operationId": "skland_checkin_now_api_skland_checkin_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckinResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckinLogOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "game_code": {
            "type": "string",
            "title": "Game Code"
          },
          "game_name": {
            "type": "string",
            "title": "Game Name"
          },
          "role_uid": {
            "type": "string",
            "title": "Role Uid"
          },
          "role_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Name"
          },
          "channel_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "status_label": {
            "type": "string",
            "title": "Status Label",
            "default": ""
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "awards_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awards Text"
          },
          "checkin_date": {
            "type": "string",
            "title": "Checkin Date"
          },
          "checked_at": {
            "type": "string",
            "format": "date-time",
            "title": "Checked At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "game_code",
          "game_name",
          "role_uid",
          "status",
          "checkin_date",
          "checked_at"
        ],
        "title": "CheckinLogOut",
        "description": "签到记录公共结构（森空岛 / 塔吉多等共用）。"
      },
...
      "CheckinResponse": {
        "properties": {
          "skipped": {
            "type": "boolean",
            "title": "Skipped",
            "default": false
          },
          "ok": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ok"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/CheckinResultItem"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "summary"
        ],
        "title": "CheckinResponse"
      },
      "CheckinResultItem": {
        "properties": {
          "game_code": {
            "type": "string",
            "title": "Game Code"
          },
          "game_name": {
            "type": "string",
            "title": "Game Name"
          },
          "role_uid": {
            "type": "string",
            "title": "Role Uid"
          },
          "role_name": {
            "type": "string",
            "title": "Role Name"
          },
          "channel_name": {
            "type": "string",
            "title": "Channel Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "status_label": {
            "type": "string",
            "title": "Status Label",
            "default": ""
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "awards_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awards Text"
          },
          "extra_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra Text"
          }
        },
        "type": "object",
        "required": [
          "game_code",
          "game_name",
          "role_uid",
          "role_name",
          "channel_name",
          "status",
          "message"
        ],
        "title": "CheckinResultItem"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SklandBindPasswordRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "maxLength": 32,
            "minLength": 5,
            "title": "Phone"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "password"
        ],
        "title": "SklandBindPasswordRequest"
      },
      "SklandBindRequest": {
        "properties": {
          "token": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 8,
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "SklandBindRequest"
      },
      "SklandBindSmsRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "maxLength": 32,
            "minLength": 5,
            "title": "Phone"
          },
          "code": {
            "type": "string",
            "maxLength": 16,
            "minLength": 4,
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "code"
        ],
        "title": "SklandBindSmsRequest"
      },
      "SklandBindSmsSendRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "maxLength": 32,
            "minLength": 5,
            "title": "Phone"
          }
        },
        "type": "object",
        "required": [
          "phone"
        ],
        "title": "SklandBindSmsSendRequest"
      },
      "SklandBindSmsSendResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "验证码已发送"
          }
        },
        "type": "object",
        "title": "SklandBindSmsSendResponse"
      },
      "SklandBindUpdate": {
        "properties": {
          "auto_checkin": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Checkin"
          },
          "checkin_hour": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 23.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkin Hour"
          },
          "checkin_minute": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 59.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkin Minute"
          }
        },
        "type": "object",
        "title": "SklandBindUpdate"
      },
      "SklandQrPollResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "bound": {
            "type": "boolean",
            "title": "Bound",
            "default": false
          },
          "auto_checkin": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Checkin"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/SklandRoleOut"
            },
            "type": "array",
            "title": "Roles"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "title": "SklandQrPollResponse"
      },
      "SklandQrStartResponse": {
        "properties": {
          "scan_id": {
            "type": "string",
            "title": "Scan Id"
          },
          "scan_url": {
            "type": "string",
            "title": "Scan Url"
          },
          "qr_image": {
            "type": "string",
            "title": "Qr Image"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "default": 180
          }
        },
        "type": "object",
        "required": [
          "scan_id",
          "scan_url",
          "qr_image"
        ],
        "title": "SklandQrStartResponse"
      },
      "SklandRoleOut": {
        "properties": {
          "game_code": {
            "type": "string",
            "title": "Game Code"
          },
          "game_name": {
            "type": "string",
            "title": "Game Name"
          },
          "uid": {
            "type": "string",
            "title": "Uid"
          },
          "role_name": {
            "type": "string",
            "title": "Role Name"
          },
          "channel_name": {
            "type": "string",
            "title": "Channel Name"
          }
        },
        "type": "object",
        "required": [
          "game_code",
          "game_name",
          "uid",
          "role_name",
          "channel_name"
        ],
        "title": "SklandRoleOut"
      },
      "SklandStatusOut": {
        "properties": {
          "bound": {
            "type": "boolean",
            "title": "Bound"
          },
          "auto_checkin": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Checkin"
          },
          "checkin_hour": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkin Hour"
          },
          "checkin_minute": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkin Minute"
          },
          "bound_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bound At"
          },
          "last_checkin_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Checkin At"
          },
          "last_checkin_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Checkin Date"
          },
          "last_checkin_ok": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Checkin Ok"
          },
          "last_checkin_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Checkin Summary"
          },
          "token_ok": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Ok"
          },
          "token_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Error"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/SklandRoleOut"
            },
            "type": "array",
            "title": "Roles"
          },
          "today_results": {
            "items": {
              "$ref": "#/components/schemas/CheckinResultItem"
            },
            "type": "array",
            "title": "Today Results"
          },
          "today_logs": {
            "items": {
              "$ref": "#/components/schemas/CheckinLogOut"
            },
            "type": "array",
            "title": "Today Logs"
          }
        },
        "type": "object",
        "required": [
          "bound"
        ],
        "title": "SklandStatusOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}